+++ src/Faker/Internal/Types.hs
@@ -137,6 +137,7 @@ data SourceData
   | Restaurant
   | RickAndMorty
   | RockBand
+  | Room
   | Rupaul
   | Science
   | Seinfeld